Refactor ebill_paynet to allow for specific discount module
The payment term node in the Paynet message allows to add a discount
information in case of payment until a certain date.
As this feature does not exist in the standard implementation of Odoo,
this is extracted into another module, to give the option to use
different implementation for it.
See `ebill_paynet_account_financial_discount` for one of them.